Change location of 'container_from_file' function
