+++ clinica/pipelines/t1_linear/t1_linear_utils.py
@@ -6,36 +6,6 @@ command line tool. See here for more details:
 http://clinica.run/doc/InteractingWithClinica/
 """
 
-
-# Get containers to produce the CAPS structure
-# Warning!!! This file should be in the future in the utils package of Clinica
-def container_from_filename(bids_or_caps_filename):
-    """Extract container from BIDS or CAPS file.
-    Args:
-       bids_or_caps_filename (str): full path to BIDS or CAPS filename.
-    Returns:
-       Container path of the form "subjects/<participant_id>/<session_id>"
-    Examples:
-       >>> from clinica.utils.nipype import container_from_filename
-       >>> container_from_filename('/path/to/bids/sub-CLNC01/ses-M00/anat/sub-CLNC01_ses-M00_T1w.nii.gz')
-               'subjects/sub-CLNC01/ses-M00'
-       >>> container_from_filename('caps/subjects/sub-CLNC01/ses-M00/dwi/preprocessing/sub-CLNC01_ses-M00_preproc.nii')
-               'subjects/sub-CLNC01/ses-M00'
-    """
-
-    import os
-    import re
-    m = re.search(r'(sub-[a-zA-Z0-9]+)/(ses-[a-zA-Z0-9]+)', bids_or_caps_filename)
-    if m is None:
-        raise ValueError(
-                'Input filename is not in a BIDS or CAPS compliant format.'
-                'It does not contain the participant and session ID.'
-                )
-    subject = m.group(1)
-    session = m.group(2)
-    return os.path.join('subjects', subject, session)
-
-
 def get_substitutions_datasink(bids_file):
 
     substitutions_ls = [  # registration
